feat: rename textfield folder
