+++ src/components/StoryBook/index.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import Box from '../Box';
 import Button from '../Button';
-import TextField from '../Text field';
+import TextField from '../TextField';
 import Typography from '../Typography';
 
 const StoryBook = () => {
